requirements.txt

streamlit==1.29.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
google-generativeai==0.3.2
python-docx==0.8.11
PyPDF2==3.0.1
fake-useragent==1.4.0
selenium==4.15.2
